Add Subscription for Refresh event
Add subscription for refresh event via the menu

The need for duplicating the following code:

    @Subscribe
    public void onRefreshEvent(RefreshEvent event) {
        startLoadingData();
    }

is due to the fact that Otto currently doesn’t traverse the hierarchy
looking for the subscription methods. That means, that if we register a
child, then it will look for the subscription ONLY in the child, and
not the parent as well. Yuck.

Looks like it’s by design, due to performance:
https://github.com/square/otto/issues/26