Fix issue with multi-platform personas
Fixing issue with multi-platform personas, as reported by Mike. I feel
this is necessary to include in this PR as they should go hand in hand.

However, it’s ugly that they do share ID’s, so that we instead have to
rely on storing the positions instead. Bleh.

Link to Pivotal:
https://www.pivotaltracker.com/story/show/67605212