Don't replace fragment if the NavigationDrawer has been recreated
Most often viewed in any scrollable views, the NavigationDrawer usually
just *restores* the last viewed fragment.

However, when we’re talking about opening a new activity, and then
going back… well, that doesn’t quite cut it. The problem that occurs is
that the scrolling position is reset, so I’ve implemented a fix to make
sure that we only actually “select item from state” when needed to.